+++ src/client.js
@@ -7,7 +7,7 @@ import { render as renderDevtools } from './client/devtools';
 // dependencies of external source. these resolve via webpack aliases
 // as assigned in merge-configs.js
 import middleware from 'middleware';
-import { createForClient as createRootComponentForClient } from 'rootComponent';
+import createRootClientComponent from 'rootClientComponent';
 
 const dest = document.getElementById('content');
 
@@ -15,7 +15,7 @@ const store = createStore(middleware, window.__data);
 const devComponent = renderDevtools();
 
 // There is probably no need to be asynchronous here
-createRootComponentForClient(store, __PROVIDERS__)
+createRootClientComponent(store, __PROVIDERS__)
   .then(({ root }) => {
     ReactDOM.render(root, dest);
 
@@ -26,7 +26,7 @@ createRootComponentForClient(store, __PROVIDERS__)
       }
     }
 
-    return devComponent ? createRootComponentForClient(store, { devComponent }, __PROVIDERS__) : {};
+    return devComponent ? createRootClientComponent(store, { devComponent }, __PROVIDERS__) : {};
   })
   .then(({ root }) => {
     if (root) ReactDOM.render(root, dest);
