+++ src/client/providers/redux-async-connect.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Provider } from 'react-redux';
+import createRouter from './react-router';
+
+export default function(store, { devComponent }) {
+  const root = (
+    <Provider store={store} key="provider">
+      <div>
+        {createRouter(store)}
+        {devComponent}
+      </div>
+    </Provider>
+  );
+
+  return Promise.resolve({ root });
+}
