+++ src/server/providers/redux-async-connect.js
@@ -1,20 +1,8 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import { createForClient as createRouterForClient } from './react-router';
 import { ReduxAsyncConnect, loadOnServer } from 'redux-async-connect';
 
-export function createForClient(store, { devComponent }) {
-  const root = (
-    <Provider store={store} key="provider">
-      {createRouterForClient(store)}
-      {devComponent}
-    </Provider>
-  );
-
-  return Promise.resolve({ root });
-}
-
-export function createForServer(store, renderProps) {
+export default function(store, renderProps) {
   return new Promise((resolve, reject) => {
     loadOnServer(renderProps, store)
       .then(() => {
