+++ src/shared/create.js
@@ -1,6 +1,10 @@
 import { map } from 'lodash';
 import createLogger from 'redux-logger';
+
+// TODO: pull our react-router deps
 import { syncHistory } from 'react-router-redux';
+import { browserHistory } from 'react-router';
+import createMemoryHistory from 'react-router/lib/createMemoryHistory';
 
 import { compose as composeDevtools, listenToRouter as linkDevtoolsToRouter } from '../client/devtools';
 import { applyMiddleware, createStore } from 'redux';
@@ -19,7 +23,13 @@ function hmr(store) {
 }
 
 export default function create(providedMiddleware, data) {
-  const router = syncHistory(history);
+  let router;
+  if (__CLIENT__) {
+    router = syncHistory(browserHistory);
+  } else {
+    router = syncHistory(createMemoryHistory());
+  }
+
   const defaultMiddleware = [ router ];
 
   // backward compatibility to 2.x api expecting object for middleware instead of array:
